Updated test class cleanup to be more precise

+++ tests/test_controller.py
@@ -25,7 +25,13 @@ class Test_Controller(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls) -> None:
-        shutil.rmtree(SCENE_HIST_DIR)
+        # remove all TEST_FILE_NAME in SCENE_HIST_DIR
+        test_files = glob.glob(f'{SCENE_HIST_DIR}/{TEST_FILE_NAME}*')
+        for test_file in test_files:
+            os.unlink(test_file)
+        # if SCENE_HIST_DIR is empty, destroy it
+        if not os.listdir(SCENE_HIST_DIR):
+            shutil.rmtree(SCENE_HIST_DIR)
 
     def create_mock_scene_event(self, mock_scene_event_data):
         # Wrap the dict in a SimpleNamespace object to permit property access
